Move dialogue-by-dialogue evaluation code to a separate file
