+++ lib/index.js
@@ -24,10 +24,11 @@ const Predictor = require('./predictor');
 const {
     SentenceEvaluatorStream,
     CollectSentenceStatistics,
-
-    DialogEvaluatorStream,
-    CollectDialogStatistics
 } = require('./evaluators');
+const {
+    DialogueEvaluatorStream,
+    CollectDialogueStatistics
+} = require('./dialogue_evaluator');
 
 const BinaryPPDB = require('./binary_ppdb');
 
@@ -64,8 +65,8 @@ module.exports = {
 
     SentenceEvaluatorStream,
     CollectSentenceStatistics,
-    DialogEvaluatorStream,
-    CollectDialogStatistics,
+    DialogueEvaluatorStream,
+    CollectDialogueStatistics,
 
     // semi-unstable API
     Utils,
